WEBAPP-9: Add PDFButton
