+++ src/components/Content/Page.js
@@ -2,6 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import style from './Page.css'
+import PDFButton from './PDFButton'
 
 class Page extends React.Component {
   static propTypes = {
@@ -10,8 +11,13 @@ class Page extends React.Component {
 
   render () {
     /* We can insert our html here directly since we trust our backend cms */
-    return <div className={style.remoteContent}
-                dangerouslySetInnerHTML={{__html: (this.props.page.content)}}/>
+    return (
+      <div>
+        <div className={style.remoteContent}
+             dangerouslySetInnerHTML={{__html: (this.props.page.content)}}/>
+        <PDFButton requestType="page" parentPage={this.props.page} pages={[this.props.page]}/>
+      </div>
+    )
   }
 }
